orpha_lookup/requirements.txt
Django==3.2.4
django-choices==1.7.1
django-environ==0.4.5
django-object-actions==3.0.2
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
django-model-utils==4.1.1
requests==2.25.1
xml_stream==0.0.7
django-bulk-update-or-create==0.3.0
more-itertools==8.8.0
